Refactor panosupgradeweb models structure
Separated the 'panosupgradeweb/models.py' into multiple files for better readability and ease of maintenance. The models are now divided into their own specific files such as 'devices.py', 'jobs.py', 'profiles.py' and 'snapshots.py'. Also, updated the import paths in 'accounts/models.py'.
