+++ backend/accounts/models.py
@@ -1,4 +1,4 @@
-# accounts/models.py
+# backend/accounts/models.py
 
 from django.contrib.auth.models import AbstractUser
 from django.db import models
