+++ backend/panosupgradeweb/models/__init__.py
@@ -1,0 +1,4 @@
+from .snapshots import Snapshot, ContentVersion, License, NetworkInterface
+from .devices import Device, DeviceType
+from .jobs import Job, JobLogEntry
+from .profiles import Profile
